DPL: refactor setNewPowerLimit
it seems to make sense to separate concers between calcPowerLimit and
setNewPowerLimit. the former just calculates the desired new power
limit. the latter enforces the limits and hyteresis, then makes sure
that the inverter is setup such that the requested power is produced.

this refactor fixes problems, e.g., a limit greater than the upper limit
does not lead to the inverter being started up, which was previously the
case. also, the lower power limit is not setup twice if the requested
power limit is below the lower limit.

the hysteresis check now compares the requested power limit with
the last set power limit, rather than with the AC power output of the
inverter. previously, the inverter would receive a new power limit
even if the last requested power limit was well within the hysteresis.
the inverter shall be left alone while the current power limit is
suitable and the inverter is still trying to adjust its actual power
output to the current power limit.

the new function commitPowerLimit shall reduce code duplication and
shall make sure that all desired actions are taken every time a new
limit was requested. in particular, this includes saving the last
requested power limit and the timestamp. the function also takes care
of starting up and shutting down the inverter power production.