JK BMS: fully decode "read all" response
this introduces a container for all values reported by the BMS,
identified by labels (enum values). values can be added using a
label and can be retrieved using a label. traits associate a particular
type for the values with each label.

decode the BMS response and throw all values into that container.