Feature: DPL: switch to allow ignoring SoC values
unfortunately, the battery SoC values reported by battery BMSs are
unreliable for some users. it offers great advantages to connect
OpenDTU-OnBattery to a BMS (MQTT publishing of values, Home Assistent
integration, etc.), but previously the users were then forced to
configure the DPL by SoC values.

this change allows to configure the DPL such that SoC values are
ignored. instead, the voltage limits are used to make DPL decisions.

the SoC related setting are hidden from the DPL settings view if SoC
values are configured to be ignored.

closes #654.