Feature: huawei grid charger: support serial CAN controller
this changeset adds support for talking to the Huawei grid charger using
a CAN transceiver like the VP230 on the OpenDTU Fusion CAN/Iso shield.
we need the ESP32's TWAI controller for that, so this feature conflicts
with the Pylontech (and all other battery providers using CAN) battery
implementation.

the user selects which hardware interface to use in the web UI/config.