huawei grid charger: avoid reboot on settings change
MCP215: make event-driven. in particular, this allows us to sleep for
some time in the loop task instead of only yield()ing, which effectively
keeps the core busy all the time because of this task. the newly added
ISR now wakes the loop task once the IRQ pin goes low.

use unique_ptr<> to manage the lifetime of the MCP2515 CAN lib object as
well as the respective SPI host object.