Fix: Huawei grid charger with MCP2515: avoid missing out on CAN messages
we really need to be quick in retrieving CAN messages when the MCP2515
signals its interrupt. we therefore increase the hardware interface
task's priority, so it is even higher priority than the Async TCP task.
more importantly, we need to ask the scheduler to preempt a low-priority
task and schedule the hardware interface task immediately, so we can
retrieve the CAN messages in time.