Fix: Huawei grid charger: known last value ends transaction
the previous implementaton by MalteSchm relied that the 0x81 output
current value was always the last one to be received. afterwards the
data would be pushed upwards the stack. I did not like that, as I think
that the values can potentially arrive in arbitrary order, as they are
encoded with a type for a reason. in practice, that is probably not the
case.

waiting to push the collected values up the stack until a new request is
sent has a significant drawback: once the values arrive at higher
software levels, they are already <poll interval> seconds old.

for that reason, I decided to restore the previous logic, where the
output current value triggers the completion of collecting values.
however, as a fall-back, the cached values are also pushed in case there
are still some when we send a new request.