limit amount of concurrent web console clients
pushing log output through a websocket to a connected client is
extremely useful, but also quite memory hungry. in order to prevent
ourselves from overwhelming out memory, we only allow two concurrent
websocket clients. the previous use of cleanupClients() would permit
a default number of clients, which is 8 for ESP32. that's too many.