R48xx: process CAN ID 0x1081407E
there is a final CAN message in the response to a "data" request, which
has the last bit in the CAN ID reset. this message clearly belongs to
the answer. it contains the value for "label" 0x83.

this is similar to the board properties answer, which consists of a
whole bunch of replies, and the last one has the LSB of the CAN ID
reset.