R48xx: gracefully handle value IDs starting with 0x31
usually, during normal operation, the value IDs in replies to the "data"
request, start with 0x01. however, in specific circumstances, the value
IDs start with 0x31 instead. this seems to happen when the PSU is
"busy", e.g., when powering up, and when shortening or opening the slot
detect pins. the values of these messages seem perfectly fine, so we
process them as usual, even though we don't know what it means that
these two bits are set in the value ID.