Feature: R48xx: support for input limit command
WARNING: this command does not work for all units!
WARNING: this command may not work as you would expect!

if the command is supported, it will limit the input current rather then
the output current. my R4850G2 unit from 2013 does NOT support it, but
the same unit from 2016 does. my R4830G1 from 2017 does as well. other
people made the same observation. this seems to be a firmware feature.

in theory, if the output voltage changes, the output current should be
re-adjusted by the rectifier automatically. my R4850G2 which supports
this feature, seems NOT to re-adjust the output voltage. other people
tested the feature with a different result:
    https://github.com/mjpalmowski/CAN-BUS-control-R4875G1-with-ESPHome-and-MQTT/discussions/4#discussioncomment-11977388