Feature: R48xx: process device's row and slot number
the slot detect pins are usually used to let the device know where in an
array of devices it is located, most certainly because it can then tell
the controller.

in the context of OpenDTU-OnBattery, users are inclined to use these
signals to enable or disable the PSU. that is a hack. still, the row and
slot number is now processed and shown in the web UI, so users can know
whether or not their slot detect hack is active or not.