Fix: ignore non-eligible inverters in all calculcations
non-eligible inverters don't participate in achieving a new target
output value in a particular DPL loop. however, we did not ignore them
in calculations leading up to setting new limits. unreachable inverters
were hence included in calculations with their last reported AC output
value, which can be significant and is always ~10W for solar-powered
inverters which went to sleep for the night.

instead of trying to calculate the household consumption, which is then
used to calculate the desired total output of all DPL-governed
inverters, we now enter a new DPL round with a target value based on the
power meter reading and eligible inverters alone: the eligible inverters
shall produce whatever they produce right now, but more or less the
value that the power meter reads. this should be quite robust, and in
particular, intermittent RF issues with inverters are not causing issues
any more. if any inverter is non-eligible for any reason, it just does
not participate *at all* in any DPL-related calculations. however, the
DPL now matches the difference using the inverter which *are*
reachable/eligible.