Fix: do not retry limit commands
once a limit command completes, we forget about it, even if it failed.
this is a good choice since the limit value might be outdated once it
timed out. the DPL shall calculate a new limit, which is then sent to
the inverter instead.

we still retry power commands to shut down or wake up inverters as it is
unlikely that the respective inverter shall have a different power state
in the next DPL loop.