Fix: cannot un-govern solar-powered inverters
the standby() method of battery- and smart-buffer-powered inverters only
sets the inverter to actual standby. reaching that state can be checked
for, so the update() method returns false (read: not busy) immediately
if the inverter is already in standby.

on the other hand, solar-powered inverters update their limit to the
configured minimum limit. testing for "is that limit already set" is not
implemented and shall not be implemented. instead, we record a timestamp
when starting the update, send the limit, and wait for it to complete
before the update() method returns false, i.e., it never returns false
on the first call.

DPL's reloadConfig() method relied on update() returning false
immediately at some point to complete reloading the config. hence, that
did not work when removing a solar-powered inverter from the list of
governed inverters.

a new concept is introduced in this changeset, getting rid of different
convoluted ways to handle that inverters should be shut down and
abandoned by the DPL. we now keep a list of retirees, i.e., inverters
that should be shut down once by the DPL and then left alone. once an
inverter is to be retired, it moves to that list. in updateInverters()
we take care of our retirees by calling the new retire() method. it will
return false eventually, at which point we remove the inverter from the
list of retirees and are done with it.