DataPointContainer::add() shall update existing data
we want to be able to throw *updated* data points into an existing data
point container using the add() method, overwriting existing data points
for that label. since we use emplace(), an existing element in the
container will not be modified. we now remove the data point using that
label beforehand.