Fix: ignore stats of non-eligible inverters
as non-eligible inverters do not participate in the current DPL loop, we
can and we should ignore whether or not we have current statistics on
them. in particular this makes for unreachable inverters, which will not
send stats for how ever long they are unreachable. the DPL shall not
wait for those inverters to become reachable before continuing to do
calculations, which ignore those inverters any way.