Fix: DPL: do not rely on unreachable inverters for voltage reading
if the inverter configured to read the battery voltage from is
unreachable, we shall not use it to determine the battery voltage. the
reading could be outdated. instead, return the error value and allow the
caller to use a different data source for the battery voltage.

in case the inverter is the only data source for the battery voltage and
it is not reachable, we won't know the battery voltage and the other
inverters are shut down as a consequence (as the battery allowance will
be zero, which is turn a consequence of the battery voltage reading of
zero is always below the stop threshold).