Fix: prioritize VE.Direct network power
the DPL shall use the VE.Direct network input power to derive
(guesstimate) the available solar output. we did this before already,
but a regression was introduced with the solar charger refactoring.

we now also try to be smarted about guesstimating the output: we reduce
the network total input power by the input power of all charge
controllers we are listening to and account their battery output power
instead. the remainder of the network input power, i.e., the input power
of all networked charge controllers that we are not listening to, is
corrected for the average efficiency and added to the return value.