Improvement: add pr_artifacts_comment step to build.yml for pull-requests
This step will post a comment on a PR with links to the firmware
artifacts created by the build action of this PR. The comment will be
updated when a new build was triggered.

only artifacts of successful build runs can and will be linked. if the
PR is updated but does not build, users should have at least a hint that
the updated changes are not part of the PR build linked in the
respective comment. add the commit SHA the artifacts were built from.
explain that these links are ZIP files.