Fix: split artifacts comment in two workflows
This is needed because a action run from a pull-requests coming from a fork does not have the required permissions to add and modify a comment.
By splitting the workflows in two we can generate and store the comment data within the pull-request action and let another workflow run afterwards that has elevated permissions.

Two different actions are used because each runs in a different context and requires different permissions. See the following article for more details. https://securitylab.github.com/resources/github-actions-preventing-pwn-requests/