Fix: DPL: always re-compute and save full solar-passthrough state
we previously failed to update the respective variable if full
solar-passthrough was not active because regular solar-passthrough was
disabled. this is a rare case, but still wrong. also, it is simply not
desirable to do the respective calculations each time the function is
called. instead, very similar to how we calculate the battery usage
allowance and restriction, we now calculate the full solar-passthrough
state in each DPL loop.