add LogHelper lib: DTU_LOG macros and dumpBytes()
this library will help us format information and logging it using the
ESP_LOG macros.

the DTU_LOG macros are wrappers around the ESP_LOG macros, expecting
SUBTAG to be defined and usin that to inject said SUBTAG into the log
message. this is useful for OpenDTU-OnBattery, as we don't want to use
distinctive tags for all providers of a module. yet, we still want to
be able to see immediately which provider or submodule produced a log
messages. example: TAG=battery and SUBTAG=Pytes or SUBTAG=Pylontech.

the first function we need is one that dumps a range of Bytes as hex.
since we need to log one line at a time when using ESP_LOG macros, this
is a little more complicated than before. we use this opportunity to
create a function that also adds the ASCII representaion of the Bytes to
the log, similar as to how the hexdump utility on GNU/Linux does it.

the LogHelper is implemented as a lib since we also want to use it from
the VE.Direct library in particular.