Fix: DPL: inverters disconnected from the grid are non-eligible
this is important in order to avoid trying to start inverters which are
not connected to the grid. starting them will fail. consecutive failure
to start the inverter will eventually lead to a reboot. this has to be
avoided. as inverters disconnected from the grid (for whatever reason)
cannot participate in producing power to cover the household
consumption, it makes sense in general to ignore these inverters.