Initial CAN Driver for JK BMS
The driver was cloned from the pylontech driver with some modifications for the hardware interface.
The CAN Interface can be configured for 250 kBits CAN Speed by adding the statement "can_type": 2 to the pin description in the hardware config.

"battery": {
    "rx": 13,
    "tx": 14,
    "can_type": 2
}

sets the speed to 250 kBits. Any other value will set it to 500 kBits which was previously the default.