undef TAG only in source files
we must not un-define TAG in a header file which is (possibly) included
before other header files, as those still might want to use the ESP-IDF
LOG* macros, which rely on TAG being a symbol. instead, we #undef TAG
right before we set our own symbol (a static const char*).